+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # A VENIR
 
+- [ ] pyweatherfr -g 89 2.15 --world corriger
+
 - [ ] integrer test sur heure soleil si ==3 si lumineux et ou petite pluie
 - [ ] comparer boundingbox openstreetmap pour doublons
 - [ ] ajouter fleche vent
@@ -11,6 +13,12 @@
 - [ ] ajouter log version and co pour issue : https://github.com/thib1984/ytdlmusic/blob/main/ytdlmusic/print.py
 - [ ] json formater pour le log/debug : https://www.freecodecamp.org/news/how-to-pretty-print-json-in-python/
 - [ ] fichier log dans le dossier config pour remplacer le mode debug avec dates and co
+
+# 5.0.1
+
+- [x] affiche un warning si ville hors france
+- [x] précise l'erreur si ville non trouvée sans --world
+- [x] remonte les villes en francais
 
 # 5.0.0